# docs/developer-portal/openapi.yaml
openapi: 3.1.0
info:
  title: VaaS API — Vehicle as a Service
  version: "0.1.0"
  description: |
    APIs de domínio do VaaS (Vehicle as a Service) — Bradesco Open Platform.

    Cobre a jornada de checkout veicular: consulta de veículo, cálculo de
    tributos por UF, simulação de financiamento, gate de política, execução
    de pagamento e agendamento de emplacamento.

    > Todos os dados retornados pelo ambiente de sandbox são **sintéticos**.
    > Ver "Sandbox & Dados Sintéticos" neste portal.

    Cada operação abaixo corresponde a uma ferramenta MCP do catálogo do
    agente — ver "Catálogo de Ferramentas MCP".
  contact:
    name: Open Platform — APIs, BaaS & Embedded Finance
    email: open-platform@bradesco.com.br
servers:
  - url: https://sandbox.api.bradesco.com.br/vaas/v1
    description: Sandbox (dados sintéticos)
  - url: http://localhost:8080/v1
    description: Ambiente local (projeto de exemplo)
security:
  - ApiKeyAuth: []
paths:
  /veiculos/{placa}:
    get:
      operationId: consultarVeiculo
      summary: Consulta dados do veículo
      tags: [Veículo]
      parameters:
        - name: placa
          in: path
          required: true
          schema: { type: string, example: ABC1D23 }
        - name: uf
          in: query
          required: true
          schema: { type: string, example: SP }
      responses:
        "200":
          description: Veículo encontrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VeiculoResponse" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    VeiculoResponse:
      type: object
      properties:
        placa: { type: string, example: ABC1D23 }
        uf: { type: string, example: SP }
        modelo: { type: string, example: "Volkswagen Polo 2023" }
        valorFipe: { type: number, example: 72400 }
...